+++ tasks/resolve_reference.rb
@@ -9,8 +9,13 @@ class ProxmoxInventory < TaskHelper
   include RubyPluginHelper
 
   attr_accessor :client
+  attr_accessor :template
+  attr_accessor :node_dns
 
-  @@node_dns = {}
+  def initialize
+    super()
+    @node_dns = {}
+  end
 
   def convert_key_value_string(value)
     value.split(',').map { |pair|
@@ -22,51 +27,78 @@ class ProxmoxInventory < TaskHelper
     }.to_h
   end
 
-  def build_agent(resource)
-    net_conf = (@client["nodes/#{resource[:node]}/#{resource[:id]}/agent/network-get-interfaces"].get)[:result]
-    net_conf.delete_if { |x| x[:name] =~ %r{^lo} }.map do |x|
-      x.delete(:statistics)
-      x['hwaddr'] = x.delete(:"hardware-address")
-      x[:"ip-addresses"].delete_if { |b| b[:"ip-address-type"] == 'ipv6' }
-      unless x[:"ip-addresses"].empty?
-        x['ip'] = x.delete(:"ip-addresses")[0][:"ip-address"]
-      end
-      x['name'] = x.delete(:name)
+  # Resolve and format qemu netX interfaces to look similar to lxc
+  # QEMU
+  # :net=>[{"virtio"=>"9E:5C:5F:5F:D0:1B", "bridge"=>"vmbr1"}]
+  # LXC
+  # :net=>[{"name"=>"eth0", "bridge"=>"vmbr1", "gw"=>"192.168.47.254", "hwaddr"=>"FA:77:1B:D7:3C:E2", "ip"=>"192.168.47.23", "type"=>"veth"}]
+  def resolve_qemu_network(resource, config)
+    unless config.key?(:net)
+      return
     end
-    { 'net' => net_conf }
+
+    config[:net].each.map do |value|
+      device_type = value.select { |k, v| k if v.match?(%r{^(([A-Za-f\d]{2}):?){6}}) }.keys[0]
+      value['hwaddr'] = value.delete(device_type).upcase
+      value.merge!({ 'type' => device_type.to_s })
+    end
+
+    begin
+      agent_network = (@client["nodes/#{resource[:node]}/#{resource[:id]}/agent/network-get-interfaces"].get)[:result]
+    rescue ProxmoxAPI::ApiException
+      # no agent running?
+      return
+    end
+
+    begin
+      config[:net].each.map do |value|
+        agent_net_intf = agent_network.find do |ani|
+          # rjust to work-around osx agent bug when hwaddr starts with zero
+          ani.key?(:'hardware-address') && ani[:"hardware-address"].rjust(17, '0').casecmp(value['hwaddr']).zero?
+        end
+        next unless agent_net_intf.class == Hash
+        unless agent_net_intf[:'ip-addresses'].empty?
+          agent_net_intf[:'ip-addresses'].delete_if { |b| b[:'ip-address-type'] == 'ipv6' }
+          unless agent_net_intf[:"ip-addresses"].empty?
+            value['ip'] = agent_net_intf[:'ip-addresses'][0][:'ip-address']
+          end
+        end
+        value['name'] = agent_net_intf[:name]
+      end
+    rescue StandardError => e
+      raise TaskHelper::Error.new("#{e.backtrace[0]} #{e.message} #{agent_network}", 'bolt-plugin/validation-error')
+    end
   end
 
   def build_data(resource)
     config = @client["nodes/#{resource[:node]}/#{resource[:id]}/config?current=1"].get
 
-    if config.key?(:agent) && config[:agent].start_with?('1')
-      begin
-        config[:agent] = build_agent(resource)
-      rescue
-        config[:agent] = nil
-      end
-    end
-
     config.keys.grep(%r{^(ipconfig|net|mp|unused)\d+}).each do |v|
       config[v.to_s.gsub!(%r{\d+}, '').to_sym] = []
     end
-
-    unless config.key?(:hostname)
-      config[:fqdn] = "#{config[:name]}"
-    else
-      config[:fqdn] = "#{config[:hostname]}"
-    end
-    unless config.key?(:searchdomain)
-      config[:fqdn] += ".#{get_node_dns(resource[:node])[:search]}"
-    else
-      config[:fqdn] += ".#{config[:searchdomain]}"
-    end
-
-    config.each { |k, v|
+    config.each do |k, v|
       if %r{^(?<index>ipconfig|net|mp|unused)(?<count>\d+)} =~ k.to_s
         config[index.to_sym][count.to_i] = convert_key_value_string(v)
       end
-    }.merge(resource)
+    end
+
+    if config.key?(:net) && config.key?(:agent) && config[:agent].start_with?('1')
+      resolve_qemu_network(resource, config)
+    end
+
+    config[:fqdn] = if config.key?(:hostname)
+                      config[:hostname].to_s
+                    else
+                      config[:name].to_s
+                    end
+
+    config[:fqdn] += if config.key?(:searchdomain)
+                       ".#{config[:searchdomain]}"
+                     else
+                       ".#{get_node_dns(resource[:node])[:search]}"
+                     end
+
+    config.merge(resource)
   end
 
   def client_config(opts)
@@ -103,21 +135,21 @@ class ProxmoxInventory < TaskHelper
   end
 
   def get_node_dns(node)
-    unless @@node_dns.key?('search')
-      @@node_dns = @client["nodes/#{node}/dns"].get
+    unless @node_dns.key?('search')
+      @node_dns = @client["nodes/#{node}/dns"].get
     end
-    @@node_dns
+    @node_dns
   end
 
   def filter_targets(targets)
     targets.delete_if do |t|
-      t[:type] == 'qemu' && t[:agent].class != Hash
+      !t.key?(:net) || t[:net].class != Array || !t[:net][0].key?('ip')
     end
   end
 
   def resolve_reference(opts)
-    template = opts.delete(:target_mapping) || {}
-    unless template.key?(:uri) || template.key?(:name)
+    @template = opts.delete(:target_mapping) || {}
+    unless @template.key?(:uri) || @template.key?(:name)
       msg = "You must provide a 'name' or 'uri' in 'target_mapping' for the Proxmox plugin"
       raise TaskHelper::Error.new(msg, 'bolt-plugin/validation-error')
     end
@@ -125,8 +157,8 @@ class ProxmoxInventory < TaskHelper
     build_client(opts)
 
     # Retrieve a list resources from the cluster
-    resources = @client.cluster.resources.get.select do |res|
-      res[:node] if res[:type] == opts[:type] && ['running'].include?(res[:status])
+    resources = @client['cluster/resources?type=vm'].get.select do |res|
+      res[:node] if (res[:type] == opts[:type] || opts[:type] == 'all') && ['running'].include?(res[:status])
     end
 
     # Retrieve node configuration
@@ -136,7 +168,7 @@ class ProxmoxInventory < TaskHelper
 
     filter_targets(targets)
 
-    attributes = required_data(template)
+    attributes = required_data(@template)
     target_data = targets.map do |target|
       attributes.each_with_object({}) do |attr, acc|
         attr = attr.first